backend/requirements.txt
Flask
flask-cors
numpy
gunicorn
pytest-mock